+++ src/components/Menu/LoginBar/LoginBar.scss
@@ -36,10 +36,12 @@
   .login-signup-content {
     height: 45%;
     background-color: #0158a3;
+
     .login-content-container {
       height: 50%;
       padding: 30px 60px;
       border-bottom: 1px solid white;
+
       .close-login-btn {
         position: relative;
         margin-left: 355px;
@@ -60,6 +62,7 @@
         .to-login-page-btn {
           width: 80px;
           font-size: 15px;
+          font-weight: bold;
           letter-spacing: 0.1em;
           border: none;
           border-radius: 20px;
@@ -74,11 +77,13 @@
       .signup-content-wrapper {
         display: flex;
         border-bottom: 1px solid white;
+
         .signup-content {
           height: 25%;
           width: 70%;
           padding: 20px 50px;
           color: white;
+
           .signup-message {
             font-size: 17px;
           }
@@ -101,12 +106,12 @@
       .signup-icon {
         font-size: 40px;
       }
-      .signup-content-container {
-      }
     }
   }
+
   .other-service-content {
     padding: 50px;
+
     .other-service-items {
       padding-bottom: 30px;
       cursor: pointer;
