+++ src/components/Nav.js
@@ -1,0 +1,15 @@
+import React from 'react';
+import TopBar from './TopBar/TopBar';
+import NavBar from './NavBar/NavBar';
+import './Nav.scss';
+
+const Nav = () => {
+  return (
+    <>
+      <TopBar />
+      <NavBar />
+    </>
+  );
+};
+
+export default Nav;
