+++ src/components/TopBar/TopBar.scss
@@ -3,9 +3,11 @@
   justify-content: space-between;
   padding: 10px 460px;
   background-color: black;
+
   .message-content {
     cursor: pointer;
     color: white;
+
     .message-text {
       font-size: 13px;
       margin-left: 10px;
